refactor: extract grid column and group logic to separate mixins
