+++ packages/grid/src/vaadin-grid-column-group-mixin.d.ts
@@ -1,0 +1,19 @@
+/**
+ * @license
+ * Copyright (c) 2016 - 2023 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import type { ColumnBaseMixinClass } from './vaadin-grid-column.js';
+
+export interface GridColumnGroupMixin<TItem> extends ColumnBaseMixinClass<TItem> {
+  /**
+   * Flex grow ratio for the column group as the sum of the ratios of its child columns.
+   * @attr {number} flex-grow
+   */
+  readonly flexGrow: number | null | undefined;
+
+  /**
+   * Width of the column group as the sum of the widths of its child columns.
+   */
+  readonly width: string | null | undefined;
+}
